Use size_t for location index instead of uint16_t
Usually, std::vector expects parameters of type size_t. I'm not going to go
into the case where it's better to use size_type for now, but probably later.

+++ src/player.hpp
@@ -11,11 +11,11 @@ public:
 	Player(std::string name, uint16_t currentHealth,
 		uint16_t maxHealth, uint16_t level, uint16_t experience,
 		uint16_t attack, uint16_t defense, uint16_t expRequired,
-		uint16_t location);
+		size_t location);
 	virtual ~Player() override;
 
 	uint16_t getExpRequired() const noexcept;
-	uint16_t getLocation() const noexcept;
+	size_t getLocation() const noexcept;
 
 	// gain an amount of experience, returning the amount of levels gained
 	bool gainExperience(uint16_t experience);
@@ -31,7 +31,7 @@ private:
 	// experience required for the next level
 	uint16_t expRequired;
 	// location id
-	uint16_t location;
+	size_t location;
 };
 
 #endif // PLAYER_HPP
